Add delegate method for column NSFormatters
This adds a delegate method to fetch an NSFormatter for each column. If a
formatter is returned, the behaviour will allow the validation and formatting of
the strings entered by the user, or of the objectValue returned from the
existing delegate method.

Note that for NSNumberFormatters in particular, you'll probably want to set
lenient to YES.

There may need to be a delegate method added for users of this project to be
able to handle invalid input and present the proper UI to the user.