WIP on allowing cells to be edited with a popup menu
There is currently a delegate method to provide whether a cell has "available
obejct values" (what will be presented as menu items in a popup menu). This can
be used to return the values or nil (for "edit as text"). Cells with available
values will display with arrows as an affordance to the user.

Todo:

- [ ] Popup cells should allow clicking the arrows once to show the popup menu
- [ ] Cancelling the popup menu should reset the editedColumn/Row ivars so that
  the cell gets redrawn (cells at those ivar coordinates are skipped in
  drawRect:)